fixed repeated local coins

+++ app.js
@@ -1,89 +1,85 @@
 //? APP.JS
-console.log("***** app *****");
+      console.log("***** app *****");
 
-const tBody = document.querySelector("tBody");
-const searchBtn = document.getElementById("search-button");
-const searchInput = document.getElementById("searchInput");
-const modalContent = document.querySelector(".modal-content");
+      const tBody = document.querySelector("tBody");
+      const searchBtn = document.getElementById("search-button");
+      const searchInput = document.getElementById("searchInput");
+      const modalContent = document.querySelector(".modal-content");
+      const savedCyrptoBtn = document.getElementById("savedCyrptoBtn");
+      const savedBody = document.querySelector(".savedBody");
+      const myModal = new bootstrap.Modal('#staticBackdrop')
+      const newDeleteBtn = document.createElement("button")
 
-const savedCyrptoBtn = document.getElementById("savedCyrptoBtn");
-const savedBody = document.querySelector(".savedBody");
-
-const myModal = new bootstrap.Modal('#staticBackdrop')
-
-// const newSaveBtn = document.getElementById("saveBtn");
-// create new delete btn
-const newDeleteBtn = document.createElement("button")
-newDeleteBtn.id = "deleteBtn";
-newDeleteBtn.className = "btn btn-primary";
-newDeleteBtn.innerText = "Delete All";
+      newDeleteBtn.id = "deleteBtn";
+      newDeleteBtn.className = "btn btn-primary";
+      newDeleteBtn.innerText = "Delete All";
 
 
-
-
-
-
-
-// https://api.coinranking.com/v2/search-suggestions?query=avax
-
-let getCoins = []
-let setLocalCoins =[]
+      let getCoins = []
+      let setLocalCoins =[]
 
 const getCripto = async () => {
-  try {
-    const response = await fetch(`https://api.coinranking.com/v2/coins`);
-    if (!response.ok) {
-      throw new Error(`Sth went wrong: ${response.status}`);
-    }
-    const data = await response.json();
-    console.log(data.data.coins);
+        try {
+            const response = await fetch(`https://api.coinranking.com/v2/coins`);
+            if (!response.ok) {
+                throw new Error(`Sth went wrong: ${response.status}`);
+            }
+            const data = await response.json();
+            console.log(data.data.coins);
 
-    showCoins(data.data.coins);
-    getCoins = data.data.coins;
- 
-  } catch (error) {
-    if (error.status === 429) {
-      console.log("Too Many Requests. Waiting for a while...");
-      await new Promise((resolve) => setTimeout(resolve, 5000));
-    } // 5 saniye bekleyin (örnek)
-    console.log(error);
-  }
+            showCoins(data.data.coins);
+            getCoins = data.data.coins;
+      
+        } catch (error) {
+            if (error.status === 429) {
+                console.log("Too Many Requests. Waiting for a while...");
+                await new Promise((resolve) => setTimeout(resolve, 5000));
+            } 
+            console.log(error);
+        }
 };
+
+
 
   const delayBetweenRequests = 1000;
   const handleApiRequests = async () => {
-  for (let i = 0; i < 3; i++) {
-    await getCripto();
-    await new Promise((resolve) => setTimeout(resolve, delayBetweenRequests));
-  }
-};
+      for (let i = 0; i < 3; i++) {
+          await getCripto();
+          await new Promise((resolve) => setTimeout(resolve, delayBetweenRequests));
+      }
+  };  
 
 function showCoins(data) {
-  data.forEach((element) => {
-    const { symbol, name, color, coinrankingUrl, iconUrl, rank, change } =
-      element;
-    const tr = document.createElement("tr");
-    tr.innerHTML = `
-            <td class="fw-bold">${name} </td>
-            <td "text-center">${rank} </td>
-            <td style=""text-center" color:${color}; font-weight:800">${symbol} </td>
-            <td "text-center">
-                <img src="${iconUrl}" alt="${name} Icon" width="30px" />
-            </td>
-            <td "text-center">${change} </td>
-            <td class="text-center">
-                    <a href="${coinrankingUrl}" target="_blank">
-                        <i class="fa-solid fa-chart-line" style="color: #4d8f5a;">
-                 
-                        </i>
-                     </a> 
-            </td>
-        `;
-    tBody.appendChild(tr);
+        data.forEach((element) => {
+          const { symbol, name, color, coinrankingUrl, iconUrl, rank, change } =
+            element;
+          const tr = document.createElement("tr");
+          tr.innerHTML = `
+                  <td class="fw-bold">${name} </td>
+                  <td "text-center">${rank} </td>
+                  <td style=""text-center" color:${color}; font-weight:800">${symbol} </td>
+                  <td "text-center">
+                      <img src="${iconUrl}" alt="${name} Icon" width="30px" />
+                  </td>
+                  <td "text-center">${change} </td>
+                  <td class="text-center">
+                          <a href="${coinrankingUrl}" target="_blank">
+                              <i class="fa-solid fa-chart-line" style="color: #4d8f5a;">
+                      
+                              </i>
+                          </a> 
+                  </td>
+              `;
+          tBody.appendChild(tr);
   });
 }
 
 handleApiRequests();
+
+   // Coin'in daha önce eklenip eklenmediğini kontrol eden fonksiyon
+   const isCoinAlreadyAdded = (coinName) => {
+    return setLocalCoins.some((addedCoin) => addedCoin.name === coinName);  
+  }
 
 const showModal = (coins) => {
 
@@ -110,10 +106,8 @@ const showModal = (coins) => {
       rank,
       change,
     } = foundCoins[0];
-    setLocalCoins.push(foundCoins[0]);
 
-
-
+    // setLocalCoins.push(foundCoins[0]);
     modalContent.innerHTML = `
               <div class="modal-header">
               <h1 class="modal-title fs-5" id="staticBackdropLabel">${name}</h1>
@@ -155,10 +149,14 @@ const showModal = (coins) => {
             
             </div>
     `;
-
-
-
-            localStorage.setItem("coins", JSON.stringify(setLocalCoins));
+         
+            // Kontrol mekanizması
+            if (!isCoinAlreadyAdded(name)) {
+              setLocalCoins.push(foundCoins[0]);
+              localStorage.setItem("coins", JSON.stringify(setLocalCoins));
+            } else {
+              console.log('Coin is already added:', name);
+            }
 
               // Yeni Save button'u oluşturan fonksiyon
               const nSaveBtn = document.getElementById("saveBtn")
@@ -166,7 +164,9 @@ const showModal = (coins) => {
               nSaveBtn.innerText = "Save";
 
               nSaveBtn.addEventListener("click", () =>{
+              myModal.hide();
               console.log("Save button is clicked");
+
               let localCoins = localStorage.getItem("coins");
 
               if (localCoins) {
@@ -180,20 +180,15 @@ const showModal = (coins) => {
                 console.log("No 'coins' key found in localStorage");
               }
               })
-
-
     myModal.show();  // Show the modal after setting its content
-  } else {
-    Swal.fire({
-      icon: "error",
-      title: "Oops...Coin not found!",
-    });
+  } 
+  else {
+      Swal.fire({
+          icon: "error",
+          title: "Oops...Coin not found!",
+      });
   }
-};
-
-
-
-
+}
 
 
 searchBtn.addEventListener("click", () => {
@@ -207,18 +202,21 @@ searchBtn.addEventListener("click", () => {
     }
     else{
       showModal(getCoins);
+      searchInput.value="";
     }
   });
 
 
+  
 
 savedCyrptoBtn.addEventListener("click", () => {
+
+    savedBody.innerHTML =``;
     let localCoins = JSON.parse(localStorage.getItem("coins"));
-    
     console.log(localCoins);
     localCoins.forEach(coin =>{
         const {name, rank, symbol, color, iconUrl, change, coinrankingUrl} = coin
-            let trNew = document.createElement("tr")
+                let  trNew = document.createElement("tr")
                 trNew.innerHTML = `
                                         <td>${name}</td>
                                         <td>${rank}</td>
@@ -239,6 +237,9 @@ savedCyrptoBtn.addEventListener("click", () => {
 
 })
 
-window.onload = ()=>{
+
+
+window.addEventListener("load", ()=>{
   searchInput.value="";
-}
+});
+
